refactorL update utils
