+++ src/dir/index.js
@@ -1,5 +1,5 @@
 export * from './cdDir.js'
-export * from './changeDir.js'
-export * from './getWorkingDir.js'
+export * from '../utils/changeDir.js'
+export * from '../utils/getWorkingDir.js'
 export * from './listDir.js'
 export * from './upDir.js'
