+++ src/utils/index.js
@@ -1,1 +1,3 @@
 export * from './handleUserInput.js'
+export * from './getWorkingDir.js'
+export * from './changeDir.js'
